requirements.txt
# NOD Lead Scraper Dependencies

# HTTP requests
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Browser automation (for JavaScript-heavy sites)
selenium>=4.15.0
webdriver-manager>=4.0.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0

# Scheduling (optional)
schedule>=1.2.0
